+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to the "BigQuery Runner" extension will be documented in this file.
 
+## [v0.0.23]
+
+### Added
+
+- Show an authentication error message.
+
+## [v0.0.22]
+
+### Changed
+
+- Fixed issue where viewer resources could not be loaded.
+
+## [v0.0.21]
+
+### Added
+
+- Add the `viewer` options to `bigqueryRunner.output.type`
+  - Rendering a well-formed table to a webview
+
 ## [v0.0.20]
 
 ### Added